Added spinner for dashboard
