+++ static/js/containers/App_test.js
@@ -8,8 +8,6 @@ import _ from 'lodash';
 import {
   CLEAR_DASHBOARD,
   CLEAR_PROFILE,
-  RECEIVE_DASHBOARD_FAILURE,
-  RECEIVE_GET_USER_PROFILE_SUCCESS,
   START_PROFILE_EDIT,
   UPDATE_PROFILE_VALIDATION,
 } from '../actions';
@@ -17,8 +15,6 @@ import {
   CLEAR_UI,
 } from '../actions/ui';
 import {
-  DASHBOARD_RESPONSE,
-  DASHBOARD_RESPONSE_ERROR,
   USER_PROFILE_RESPONSE
 } from '../constants';
 import IntegrationTestHelper from '../util/integration_test_helper';
@@ -26,7 +22,6 @@ import IntegrationTestHelper from '../util/integration_test_helper';
 describe('App', () => {
   let listenForActions, renderComponent, helper;
   let editProfileActions;
-  let dashboardErrorActions;
 
   beforeEach(() => {
     helper = new IntegrationTestHelper();
@@ -35,10 +30,6 @@ describe('App', () => {
     editProfileActions = [
       START_PROFILE_EDIT,
       UPDATE_PROFILE_VALIDATION
-    ];
-    dashboardErrorActions = [
-      RECEIVE_DASHBOARD_FAILURE,
-      RECEIVE_GET_USER_PROFILE_SUCCESS
     ];
   });
 
@@ -50,59 +41,6 @@ describe('App', () => {
     return renderComponent("/dashboard").then(([, div]) => {
       return listenForActions([CLEAR_DASHBOARD, CLEAR_PROFILE, CLEAR_UI], () => {
         ReactDOM.unmountComponentAtNode(div);
-      });
-    });
-  });
-
-  describe('dashboard errors', () => {
-    let errorString = `Sorry, we were unable to load the data necessary
-      to process your request. Please reload the page.`;
-    errorString = errorString.replace(/\s\s+/g, ' ');
-
-    it('error from the backend triggers error message in dashboard', () => {
-      helper.dashboardStub.returns(Promise.reject(DASHBOARD_RESPONSE_ERROR));
-
-      return renderComponent("/dashboard", dashboardErrorActions, false).then(([, div]) => {
-        let message = div.getElementsByClassName('alert-message')[0];
-        assert(message.textContent.indexOf(errorString) > -1);
-        assert(message.textContent.indexOf(DASHBOARD_RESPONSE_ERROR.error_code) > -1);
-        assert(message.textContent.indexOf("Additional info:") > -1);
-        assert(message.textContent.indexOf(DASHBOARD_RESPONSE_ERROR.user_message) > -1);
-      });
-    });
-
-    it('the error from the backend does not need to be complete', () => {
-      let response = _.cloneDeep(DASHBOARD_RESPONSE_ERROR);
-      delete response.user_message;
-      helper.dashboardStub.returns(Promise.reject(response));
-
-      return renderComponent("/dashboard", dashboardErrorActions, false).then(([, div]) => {
-        let message = div.getElementsByClassName('alert-message')[0];
-        assert(message.textContent.indexOf(errorString) > -1);
-        assert(message.textContent.indexOf(DASHBOARD_RESPONSE_ERROR.error_code) > -1);
-        assert.equal(message.textContent.indexOf("Additional info:"), -1);
-        assert.equal(message.textContent.indexOf(DASHBOARD_RESPONSE_ERROR.user_message), -1);
-      });
-    });
-
-    it('the error from the backend does not need to exist at all as long as there is an http error', () => {
-      helper.dashboardStub.returns(Promise.reject({}));
-
-      return renderComponent("/dashboard", dashboardErrorActions, false).then(([, div]) => {
-        let message = div.getElementsByClassName('alert-message')[0];
-        assert(message.textContent.indexOf(errorString) > -1);
-        assert.equal(message.textContent.indexOf(DASHBOARD_RESPONSE_ERROR.error_code), -1);
-        assert.equal(message.textContent.indexOf("Additional info:"), -1);
-        assert.equal(message.textContent.indexOf(DASHBOARD_RESPONSE_ERROR.user_message), -1);
-      });
-    });
-
-    it('a regular response does not show the error', () => {
-      helper.dashboardStub.returns(Promise.resolve(DASHBOARD_RESPONSE));
-
-      return renderComponent("/dashboard").then(([, div]) => {
-        let message = div.getElementsByClassName('alert-message')[0];
-        assert.equal(message, undefined);
       });
     });
   });
