+++ static/js/containers/DashboardPage.js
@@ -2,9 +2,13 @@
 /* global SETTINGS: false */
 import React from 'react';
 import { connect } from 'react-redux';
+import Loader from 'react-loader';
+
+import { FETCH_PROCESSING } from '../actions';
 import Jumbotron from '../components/Jumbotron';
 import CourseList from '../components/CourseList';
 import ErrorMessage from '../components/ErrorMessage';
+import { getPreferredName } from '../util/util';
 
 class DashboardPage extends React.Component {
   static propTypes = {
@@ -21,7 +25,8 @@ class DashboardPage extends React.Component {
       dispatch,
       profile: { profile },
     } = this.props;
-    let preferredName = profile.preferredName || SETTINGS.name;
+    const loaded = dashboard.fetchStatus !== FETCH_PROCESSING;
+    let preferredName = getPreferredName(profile);
     let errorMessage;
     let dashboardContent;
     // if there are no errors coming from the backend, simply show the dashboard
@@ -39,10 +44,10 @@ class DashboardPage extends React.Component {
     }
     return (
       <Jumbotron profile={profile} text={preferredName}>
-        <div>
+        <Loader loaded={loaded}>
           {errorMessage}
           {dashboardContent}
-        </div>
+        </Loader>
       </Jumbotron>
     );
   }
